refactor: decompose analysis.py into analysis_pkg submodules
Split the 1404-line analysis.py into three focused submodules:
- analysis_pkg/_archive.py: archive extraction, PreparedTarget, error types
- analysis_pkg/_text.py: constants, text processing, IOC/dep extraction, intel helpers
- analysis_pkg/_scanner.py: scan() entry point

analysis.py is now a thin compatibility shim that re-exports all public
symbols, so all existing imports (from skillscan.analysis import ...)
continue to work unchanged.

Fixes: ruff clean, mypy clean, all 402 tests pass
Audit finding: M-REFACTOR from AUDIT_2026-03.md
