+++ src/skillscan/analysis.py
@@ -1,1404 +1,60 @@
-from __future__ import annotations
-
-import base64
-import binascii
-import concurrent.futures
-import ipaddress
-import json
-import os
-import re
-import tarfile
-import tempfile
-import unicodedata
-import zipfile
-from dataclasses import dataclass
-from importlib import resources
-from pathlib import Path
-from typing import cast
-
-from skillscan import __version__
+# analysis.py — compatibility shim
+# The implementation has been decomposed into:
+#   skillscan/analysis_pkg/_archive.py  — archive helpers, dataclasses, prepare_target
+#   skillscan/analysis_pkg/_text.py     — text processing, IOC/dep extraction, intel helpers
+#   skillscan/analysis_pkg/_scanner.py  — scan() entry point
+#
+# All existing imports (from skillscan.analysis import ...) continue to work unchanged.
+from skillscan.analysis_pkg._archive import (
+    ArchiveFormatError,
+    ArchivePasswordError,
+    BinaryArtifact,
+    FileInventory,
+    PreparedTarget,
+    ScanError,
+    _safe_extract_7z,
+    _safe_extract_tar,
+    _safe_extract_zip,
+    _safe_extract_zst,
+    prepare_target,
+)
+from skillscan.analysis_pkg._scanner import scan
+from skillscan.analysis_pkg._text import (
+    _decode_base64_fragments,
+    _extract_actions,
+    _extract_actions_windowed,
+    _extract_iocs,
+    _find_unpinned_requirements,
+    _ip_in_cidrs,
+    _merge_user_intel,
+    _normalize_text,
+    _prepare_analysis_text,
+    iter_text_files,
+)
 from skillscan.clamav import scan_paths as clamav_scan_paths
-from skillscan.detectors.ast_flows import detect_python_ast_flows
-from skillscan.ecosystems import detect_ecosystems
-from skillscan.intel import load_store
-from skillscan.ml_detector import ml_prompt_injection_findings
-from skillscan.models import (
-    IOC,
-    Capability,
-    DependencyFinding,
-    Finding,
-    Policy,
-    ScanMetadata,
-    ScanReport,
-    Severity,
-    TriageMetadata,
-    Verdict,
-    detect_archive_format,
-    is_archive,
-)
-from skillscan.remote import RemoteFetchError, fetch_remote_target, is_url_target
-from skillscan.rules import CompiledRulePack, load_compiled_builtin_rulepack
-from skillscan.semantic_local import (
-    classify_prompt_injection_raw,
-    classify_social_engineering_raw,
-    local_prompt_injection_findings,
-    local_social_engineering_findings,
-)
 
-URL_RE = re.compile(r"https?://[^\s\"'<>]+", re.IGNORECASE)
-IP_RE = re.compile(r"\b(?:\d{1,3}\.){3}\d{1,3}\b")
-DOMAIN_RE = re.compile(r"\b[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}\b")
-ZERO_WIDTH_RE = re.compile(r"[\u200b-\u200f\u2060\ufeff]")
-B64_TOKEN_RE = re.compile(r"(?<![A-Za-z0-9+/_-])[A-Za-z0-9+/_-]{16,}={0,2}(?![A-Za-z0-9+/_=-])")
-QUOTED_B64_FRAGMENT_RE = re.compile(r"[\"']([A-Za-z0-9+/_-]{8,}={0,2})[\"']")
-RISKY_NPM_SCRIPT_RE = re.compile(
-    r"\b(curl|wget|invoke-webrequest|powershell|pwsh|bash|sh|cmd\.exe|certutil|bitsadmin|nc|ncat|netcat)\b|"
-    r"https?://",
-    re.IGNORECASE,
-)
-KNOWN_TLDS = {
-    "ai",
-    "app",
-    "biz",
-    "co",
-    "com",
-    "dev",
-    "edu",
-    "fr",
-    "gov",
-    "info",
-    "io",
-    "jp",
-    "net",
-    "org",
-    "ru",
-    "uk",
-    "us",
-    "xyz",
-}
-
-SEVERITY_SCORE = {
-    Severity.LOW: 5,
-    Severity.MEDIUM: 15,
-    Severity.HIGH: 35,
-    Severity.CRITICAL: 60,
-}
-
-IOCDB = dict[str, list[str]]
-VulnRecord = dict[str, str]
-VulnVersionMap = dict[str, VulnRecord]
-VulnPackageMap = dict[str, VulnVersionMap]
-VulnDB = dict[str, VulnPackageMap]
-
-BYTECODE_SUFFIXES = {".pyc", ".pyo"}
-SCRIPT_SUFFIXES = {
-    ".py",
-    ".sh",
-    ".bash",
-    ".rb",
-    ".js",
-    ".ts",
-    ".mjs",
-    ".cjs",
-    ".go",
-    ".rs",
-    ".pl",
-    ".ps1",
-    ".psm1",
-}
-
-# Language tag → file extensions (for multilang rule filtering)
-_LANG_EXTENSIONS: dict[str, frozenset[str]] = {
-    "javascript": frozenset({".js", ".ts", ".mjs", ".cjs"}),
-    "ruby": frozenset({".rb"}),
-    "go": frozenset({".go"}),
-    "rust": frozenset({".rs"}),
-    "python": frozenset({".py"}),
-    "shell": frozenset({".sh", ".bash", ".zsh"}),
-}
-
-# Proximity window for chain rule detection (lines)
-_CHAIN_WINDOW_LINES = 40
-
-LIBRARY_SUFFIXES = {".so", ".dll", ".dylib", ".a"}
-EXECUTABLE_SUFFIXES = {".exe", ".msi", ".com"}
-KNOWN_BINARY_SUFFIXES = {
-    ".bin",
-    ".o",
-    ".obj",
-    ".class",
-    ".jar",
-    ".wasm",
-    ".pyz",
-    ".whl",
-}
-MACHO_MAGICS = (
-    b"\xfe\xed\xfa\xce",
-    b"\xce\xfa\xed\xfe",
-    b"\xfe\xed\xfa\xcf",
-    b"\xcf\xfa\xed\xfe",
-    b"\xca\xfe\xba\xbe",
-    b"\xbe\xba\xfe\xca",
-)
-
-
-@dataclass
-class PreparedTarget:
-    root: Path
-    target_type: str
-    cleanup_dir: tempfile.TemporaryDirectory[str] | None
-    read_warnings: list[str]
-    policy_warnings: list[str]
-
-
-@dataclass
-class BinaryArtifact:
-    path: Path
-    kind: str
-    detail: str
-
-
-@dataclass
-class FileInventory:
-    text_files: list[Path]
-    binary_artifacts: list[BinaryArtifact]
-
-
-class ScanError(Exception):
-    pass
-
-
-class ArchivePasswordError(Exception):
-    """Raised when an archive is password-protected and cannot be extracted."""
-
-
-class ArchiveFormatError(Exception):
-    """Raised when an archive format is not supported by available libraries."""
-
-
-def _safe_extract_zip(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
-    total = 0
-    with zipfile.ZipFile(src) as zf:
-        infos = zf.infolist()
-        if len(infos) > max_files:
-            raise ScanError(f"Archive has too many files: {len(infos)}")
-        for info in infos:
-            name = info.filename
-            if name.startswith("/") or ".." in Path(name).parts:
-                raise ScanError(f"Unsafe archive path: {name}")
-            total += info.file_size
-            if total > max_bytes:
-                raise ScanError("Archive exceeds max bytes limit")
-            zf.extract(info, dst)
-
-
-def _safe_extract_tar(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
-    total = 0
-    with tarfile.open(src) as tf:
-        members = tf.getmembers()
-        if len(members) > max_files:
-            raise ScanError(f"Archive has too many files: {len(members)}")
-        for member in members:
-            p = Path(member.name)
-            if p.is_absolute() or ".." in p.parts:
-                raise ScanError(f"Unsafe archive path: {member.name}")
-            if member.issym() or member.islnk():
-                raise ScanError(f"Symlink/hardlink not allowed in archive: {member.name}")
-            total += member.size
-            if total > max_bytes:
-                raise ScanError("Archive exceeds max bytes limit")
-        # filter='data' was added in Python 3.12; use keyword argument only when available
-        if hasattr(tf, "extraction_filter"):
-            tf.extractall(dst, filter="data")
-        else:
-            for member in members:
-                tf.extract(member, dst)
-
-
-def _safe_extract_7z(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
-    try:
-        import py7zr  # optional extra
-    except ImportError as exc:
-        raise ArchiveFormatError("py7zr not installed; install skillscan[archives]") from exc
-    try:
-        with py7zr.SevenZipFile(src, mode="r") as sz:
-            entries = sz.list()
-            if len(entries) > max_files:
-                raise ScanError(f"Archive has too many files: {len(entries)}")
-            total = sum(e.uncompressed for e in entries if e.uncompressed)
-            if total > max_bytes:
-                raise ScanError("Archive exceeds max bytes limit")
-            for e in entries:
-                p = Path(e.filename)
-                if p.is_absolute() or ".." in p.parts:
-                    raise ScanError(f"Unsafe archive path: {e.filename}")
-            sz.extractall(path=dst)
-    except py7zr.exceptions.PasswordRequired as exc:
-        raise ArchivePasswordError("7z archive is password-protected") from exc
-
-
-def _safe_extract_rar(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
-    try:
-        import rarfile  # optional extra
-    except ImportError as exc:
-        raise ArchiveFormatError("rarfile not installed; install skillscan[archives]") from exc
-    try:
-        with rarfile.RarFile(src) as rf:
-            infos = rf.infolist()
-            if len(infos) > max_files:
-                raise ScanError(f"Archive has too many files: {len(infos)}")
-            total = sum(i.file_size for i in infos)
-            if total > max_bytes:
-                raise ScanError("Archive exceeds max bytes limit")
-            for info in infos:
-                p = Path(info.filename)
-                if p.is_absolute() or ".." in p.parts:
-                    raise ScanError(f"Unsafe archive path: {info.filename}")
-            rf.extractall(path=dst)
-    except rarfile.PasswordRequired as exc:
-        raise ArchivePasswordError("RAR archive is password-protected") from exc
-    except rarfile.NeedFirstVolume as exc:
-        raise ArchiveFormatError("Multi-volume RAR not supported") from exc
-
-
-def _safe_extract_xz(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
-    """Extract .xz (plain xz-compressed file, not tar.xz — that's handled by tarfile)."""
-    import lzma
-
-    out_name = src.stem  # strip .xz
-    out_path = dst / out_name
-    total = 0
-    with lzma.open(src, "rb") as fh:
-        while True:
-            chunk = fh.read(65536)
-            if not chunk:
-                break
-            total += len(chunk)
-            if total > max_bytes:
-                raise ScanError("Archive exceeds max bytes limit")
-            with out_path.open("ab") as out:
-                out.write(chunk)
-
-
-def _safe_extract_bz2(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
-    """Extract .bz2 (plain bzip2-compressed file, not tar.bz2 — that's handled by tarfile)."""
-    import bz2
-
-    out_name = src.stem
-    out_path = dst / out_name
-    total = 0
-    with bz2.open(src, "rb") as fh:
-        while True:
-            chunk = fh.read(65536)
-            if not chunk:
-                break
-            total += len(chunk)
-            if total > max_bytes:
-                raise ScanError("Archive exceeds max bytes limit")
-            with out_path.open("ab") as out:
-                out.write(chunk)
-
-
-def _safe_extract_zst(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
-    """Extract .zst (plain Zstandard-compressed file)."""
-    try:
-        import zstandard  # optional extra
-    except ImportError as exc:
-        raise ArchiveFormatError("zstandard not installed; install skillscan[archives]") from exc
-    out_name = src.stem
-    out_path = dst / out_name
-    total = 0
-    dctx = zstandard.ZstdDecompressor()
-    with src.open("rb") as ifh, out_path.open("wb") as ofh:
-        for chunk in dctx.read_to_iter(ifh, read_size=65536):
-            total += len(chunk)
-            if total > max_bytes:
-                raise ScanError("Archive exceeds max bytes limit")
-            ofh.write(chunk)
-
-
-def prepare_target(
-    target: Path | str,
-    policy: Policy,
-    url_max_links: int = 25,
-    url_timeout_seconds: int = 12,
-    url_same_origin_only: bool = True,
-) -> PreparedTarget:
-    if isinstance(target, str) and is_url_target(target):
-        try:
-            fetched = fetch_remote_target(
-                target,
-                max_links=url_max_links,
-                timeout_seconds=url_timeout_seconds,
-                same_origin_only=url_same_origin_only,
-            )
-        except RemoteFetchError as exc:
-            raise ScanError(str(exc)) from exc
-        return PreparedTarget(
-            root=fetched.root,
-            target_type="url",
-            cleanup_dir=fetched.cleanup_dir,
-            read_warnings=fetched.unreadable_urls,
-            policy_warnings=fetched.skipped_urls,
-        )
-
-    if isinstance(target, str):
-        target = Path(target)
-
-    if not target.exists():
-        raise ScanError(f"Target does not exist: {target}")
-    if target.is_dir():
-        return PreparedTarget(
-            root=target,
-            target_type="directory",
-            cleanup_dir=None,
-            read_warnings=[],
-            policy_warnings=[],
-        )
-    if target.is_file() and not is_archive(target):
-        tmp = tempfile.TemporaryDirectory(prefix="skillscan-")
-        dst = Path(tmp.name)
-        (dst / target.name).write_bytes(target.read_bytes())
-        return PreparedTarget(
-            root=dst,
-            target_type="file",
-            cleanup_dir=tmp,
-            read_warnings=[],
-            policy_warnings=[],
-        )
-    if target.is_file() and is_archive(target):
-        tmp = tempfile.TemporaryDirectory(prefix="skillscan-")
-        dst = Path(tmp.name)
-        fmt = detect_archive_format(target)
-        policy_warnings: list[str] = []
-        try:
-            if fmt == "zip":
-                _safe_extract_zip(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
-            elif fmt in ("gz", "bz2", "xz", "tar"):
-                # tarfile handles .tar, .tar.gz, .tar.bz2, .tar.xz transparently
-                try:
-                    _safe_extract_tar(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
-                except tarfile.ReadError:
-                    # Plain compressed file (not tar) — fall through to single-file decompressors
-                    if fmt == "xz":
-                        _safe_extract_xz(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
-                    elif fmt == "bz2":
-                        _safe_extract_bz2(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
-                    else:
-                        raise
-            elif fmt == "7z":
-                _safe_extract_7z(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
-            elif fmt == "rar":
-                _safe_extract_rar(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
-            elif fmt == "zst":
-                _safe_extract_zst(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
-            else:
-                raise ArchiveFormatError(f"Unsupported archive format: {fmt}")
-        except ArchivePasswordError:
-            # Password-protected: copy as-is and emit BIN-OPAQUE-002 during scan
-            (dst / target.name).write_bytes(target.read_bytes())
-            policy_warnings.append(f"BIN-OPAQUE-002:{target.name}")
-        except ArchiveFormatError as exc:
-            # Library not available or format not supported: copy as-is and emit BIN-OPAQUE-001
-            (dst / target.name).write_bytes(target.read_bytes())
-            policy_warnings.append(f"BIN-OPAQUE-001:{target.name}:{exc}")
-        return PreparedTarget(
-            root=dst,
-            target_type="archive",
-            cleanup_dir=tmp,
-            read_warnings=[],
-            policy_warnings=policy_warnings,
-        )
-    raise ScanError("Unsupported target type")
-
-
-def _read_head(path: Path, limit: int = 4096) -> bytes:
-    try:
-        with path.open("rb") as fh:
-            return fh.read(limit)
-    except OSError:
-        return b""
-
-
-def _classify_non_text(path: Path) -> BinaryArtifact | None:
-    suffix = path.suffix.lower()
-    if "__pycache__" in path.parts or suffix in BYTECODE_SUFFIXES:
-        return BinaryArtifact(path=path, kind="python_bytecode", detail="compiled python bytecode artifact")
-    if suffix in LIBRARY_SUFFIXES:
-        return BinaryArtifact(path=path, kind="binary_library", detail="compiled library artifact")
-    if suffix in EXECUTABLE_SUFFIXES:
-        return BinaryArtifact(path=path, kind="executable_binary", detail="known executable extension")
-
-    head = _read_head(path)
-    if not head:
-        return None
-    if head.startswith(b"\x7fELF") or head.startswith(b"MZ") or head.startswith(MACHO_MAGICS):
-        return BinaryArtifact(path=path, kind="executable_binary", detail="executable binary header")
-    if suffix in KNOWN_BINARY_SUFFIXES:
-        return BinaryArtifact(path=path, kind="binary_blob", detail="binary extension")
-    if b"\x00" in head:
-        return BinaryArtifact(path=path, kind="binary_blob", detail="contains NUL bytes")
-    return None
-
-
-def iter_text_files(
-    root: Path,
-    max_files: int,
-    max_bytes: int,
-    max_binary_artifacts: int,
-    max_binary_bytes: int,
-) -> FileInventory:
-    files: list[Path] = []
-    binary_artifacts: list[BinaryArtifact] = []
-    total_bytes = 0
-    total_files = 0
-    total_binary_bytes = 0
-    for path in root.rglob("*"):
-        if not path.is_file():
-            continue
-        try:
-            size = path.stat().st_size
-        except OSError:
-            continue
-        total_files += 1
-        total_bytes += size
-        if total_bytes > max_bytes:
-            raise ScanError("Scan size exceeded max_bytes policy limit")
-        if total_files > max_files:
-            raise ScanError("Scan file count exceeded max_files policy limit")
-        classification = _classify_non_text(path)
-        if classification is not None:
-            binary_artifacts.append(classification)
-            total_binary_bytes += size
-            if len(binary_artifacts) > max_binary_artifacts:
-                raise ScanError("Binary artifact count exceeded max_binary_artifacts policy limit")
-            if total_binary_bytes > max_binary_bytes:
-                raise ScanError("Binary artifact bytes exceeded max_binary_bytes policy limit")
-            continue
-        files.append(path)
-    return FileInventory(text_files=files, binary_artifacts=binary_artifacts)
-
-
-def _safe_read_text(path: Path) -> str:
-    try:
-        raw = path.read_bytes()
-    except OSError:
-        return ""
-    if b"\x00" in raw:
-        return ""
-    return raw.decode("utf-8", errors="ignore")
-
-
-def _normalize_text(text: str) -> str:
-    norm = unicodedata.normalize("NFKC", text)
-    norm = ZERO_WIDTH_RE.sub("", norm)
-    # Recover common defanged URL/domain forms.
-    norm = re.sub(r"\bhxxps://", "https://", norm, flags=re.IGNORECASE)
-    norm = re.sub(r"\bhxxp://", "http://", norm, flags=re.IGNORECASE)
-    norm = norm.replace("[.]", ".").replace("(.)", ".").replace("{.}", ".")
-    return norm
-
-
-def _try_decode_b64(token: str) -> str | None:
-    value = token.strip()
-    if len(value) < 16:
-        return None
-    # Normalize urlsafe base64 and missing padding.
-    value = value.replace("-", "+").replace("_", "/")
-    padding = (-len(value)) % 4
-    if padding:
-        value += "=" * padding
-    try:
-        blob = base64.b64decode(value, validate=True)
-    except (binascii.Error, ValueError):
-        return None
-    if not blob:
-        return None
-    candidate = blob.decode("utf-8", errors="ignore").strip()
-    if not candidate:
-        return None
-    printable_ratio = sum(ch.isprintable() for ch in candidate) / max(len(candidate), 1)
-    if printable_ratio < 0.9:
-        return None
-    return candidate
-
-
-def _decode_base64_fragments(text: str, max_decodes: int = 6) -> list[str]:
-    decoded: list[str] = []
-    seen: set[str] = set()
-    for token in B64_TOKEN_RE.findall(text):
-        if token in seen:
-            continue
-        seen.add(token)
-        if len(decoded) >= max_decodes:
-            break
-        candidate = _try_decode_b64(token)
-        if candidate:
-            decoded.append(candidate)
-
-    # Also try concatenated quoted fragments often used to bypass token scanners.
-    for line in text.splitlines():
-        fragments = QUOTED_B64_FRAGMENT_RE.findall(line)
-        if len(fragments) < 2:
-            continue
-        joined = "".join(fragments)
-        if joined in seen:
-            continue
-        seen.add(joined)
-        if len(decoded) >= max_decodes:
-            break
-        candidate = _try_decode_b64(joined)
-        if candidate:
-            decoded.append(candidate)
-    return decoded
-
-
-def _prepare_analysis_text(text: str) -> str:
-    norm = _normalize_text(text)
-    decoded = _decode_base64_fragments(norm)
-    if not decoded:
-        return norm
-    return f"{norm}\n\n# decoded_fragments\n" + "\n".join(decoded)
-
-
-def _extract_actions(text: str, action_patterns: dict[str, re.Pattern[str]]) -> set[str]:
-    """Return the set of action names whose patterns match anywhere in *text*.
-
-    Used for whole-file matching (backward-compatible). For chain rule detection
-    use :func:`_extract_actions_windowed` which applies a proximity constraint.
-    """
-    actions: set[str] = set()
-    for action, pattern in action_patterns.items():
-        if pattern.search(text):
-            actions.add(action)
-    return actions
-
-
-def _extract_actions_windowed(
-    text: str,
-    action_patterns: dict[str, re.Pattern[str]],
-    window_lines: int = _CHAIN_WINDOW_LINES,
-) -> list[set[str]]:
-    """Return a list of action sets, one per sliding window of *window_lines* lines.
-
-    For files shorter than *window_lines* a single whole-file window is returned,
-    preserving the original behaviour for small skill files.
-    """
-    lines = text.splitlines()
-    if len(lines) <= window_lines:
-        return [_extract_actions(text, action_patterns)]
-    windows: list[set[str]] = []
-    step = max(1, window_lines // 2)  # 50 % overlap
-    for start in range(0, len(lines), step):
-        chunk = "\n".join(lines[start : start + window_lines])
-        window_actions = _extract_actions(chunk, action_patterns)
-        if window_actions:
-            windows.append(window_actions)
-    return windows if windows else [set()]
-
-
-def _normalize_domain(value: str) -> str:
-    return value.lower().strip().strip(".,;)")
-
-
-def _parse_requirements(text: str) -> list[tuple[str, str]]:
-    deps: list[tuple[str, str]] = []
-    for line in text.splitlines():
-        line = line.strip()
-        if not line or line.startswith("#"):
-            continue
-        if "==" in line:
-            name, version = line.split("==", 1)
-            deps.append((name.strip().lower(), version.strip()))
-    return deps
-
-
-def _find_unpinned_requirements(text: str) -> list[tuple[str, str]]:
-    unpinned: list[tuple[str, str]] = []
-    for line in text.splitlines():
-        line = line.strip()
-        if not line or line.startswith("#"):
-            continue
-        if "==" in line:
-            continue
-        if any(op in line for op in (">=", "<=", "~=", ">", "<")):
-            unpinned.append((line.split()[0], line))
-    return unpinned
-
-
-def _parse_package_json(text: str) -> list[tuple[str, str]]:
-    deps: list[tuple[str, str]] = []
-    try:
-        data = json.loads(text)
-    except json.JSONDecodeError:
-        return deps
-    for section in ("dependencies", "devDependencies"):
-        values = data.get(section, {})
-        if isinstance(values, dict):
-            for name, version in values.items():
-                deps.append((name.lower(), str(version)))
-    return deps
-
-
-def _parse_package_scripts(text: str) -> dict[str, str]:
-    try:
-        data = json.loads(text)
-    except json.JSONDecodeError:
-        return {}
-    scripts = data.get("scripts", {})
-    if not isinstance(scripts, dict):
-        return {}
-    parsed: dict[str, str] = {}
-    for name, cmd in scripts.items():
-        if isinstance(name, str) and isinstance(cmd, str):
-            parsed[name.lower()] = cmd
-    return parsed
-
-
-def _is_unpinned_npm(version: str) -> bool:
-    v = version.strip().lower()
-    if not v or v == "latest":
-        return True
-    return v.startswith("^") or v.startswith("~") or "*" in v or "x" in v
-
-
-def _extract_iocs(path: Path, text: str) -> list[IOC]:
-    iocs: list[IOC] = []
-    seen: set[tuple[str, str]] = set()
-    for url in URL_RE.findall(text):
-        key = ("url", url)
-        if key not in seen:
-            seen.add(key)
-            iocs.append(IOC(value=url, kind="url", source_path=str(path)))
-    for token in IP_RE.findall(text):
-        try:
-            ipaddress.ip_address(token)
-        except ValueError:
-            continue
-        key = ("ip", token)
-        if key not in seen:
-            seen.add(key)
-            iocs.append(IOC(value=token, kind="ip", source_path=str(path)))
-    for domain in DOMAIN_RE.findall(text):
-        norm = _normalize_domain(domain)
-        if norm.startswith("http"):
-            continue
-        tld = norm.rsplit(".", 1)[-1]
-        if tld not in KNOWN_TLDS:
-            continue
-        key = ("domain", norm)
-        if key not in seen:
-            seen.add(key)
-            iocs.append(IOC(value=norm, kind="domain", source_path=str(path)))
-    return iocs
-
-
-def _load_builtin_vuln_db() -> VulnDB:
-    raw = resources.files("skillscan.data.intel").joinpath("vuln_db.json").read_text(encoding="utf-8")
-    return cast(VulnDB, json.loads(raw))
-
-
-def _load_builtin_ioc_db() -> IOCDB:
-    raw = resources.files("skillscan.data.intel").joinpath("ioc_db.json").read_text(encoding="utf-8")
-    return cast(IOCDB, json.loads(raw))
-
-
-def _merge_user_intel(ioc_db: IOCDB, vuln_db: VulnDB) -> tuple[IOCDB, VulnDB, list[str]]:
-    store = load_store()
-    sources = ["builtin:ioc_db", "builtin:vuln_db"]
-    for source in store.sources:
-        if not source.enabled:
-            continue
-        path = Path(source.path)
-        if not path.exists():
-            continue
-        try:
-            payload = json.loads(path.read_text(encoding="utf-8"))
-        except (json.JSONDecodeError, OSError, UnicodeDecodeError):
-            continue
-        if source.kind == "ioc" and isinstance(payload, dict):
-            for key in ("domains", "ips", "urls", "cidrs"):
-                values = payload.get(key, [])
-                if isinstance(values, list):
-                    ioc_db.setdefault(key, []).extend([str(v).lower() for v in values])
-        if source.kind == "vuln" and isinstance(payload, dict):
-            for ecosystem, values in payload.items():
-                eco = vuln_db.setdefault(ecosystem, {})
-                if isinstance(values, dict):
-                    eco.update(values)
-        sources.append(f"user:{source.name}")
-    for key in ("domains", "ips", "urls", "cidrs"):
-        deduped = sorted(set(ioc_db.get(key, [])))
-        ioc_db[key] = deduped
-    return ioc_db, vuln_db, sources
-
-
-def _binary_kind_template(kind: str) -> tuple[str, Severity, str, str]:
-    if kind == "executable_binary":
-        return (
-            "BIN-001",
-            Severity.HIGH,
-            "Executable binary artifact present",
-            "Review and verify binary provenance and signatures before trust or execution.",
-        )
-    if kind == "binary_library":
-        return (
-            "BIN-002",
-            Severity.MEDIUM,
-            "Compiled library artifact present",
-            "Verify shared/static library provenance and hash against trusted release artifacts.",
-        )
-    if kind == "binary_blob":
-        return (
-            "BIN-003",
-            Severity.MEDIUM,
-            "Binary blob artifact present",
-            "Inspect binary payload purpose and source; prefer transparent source artifacts when possible.",
-        )
-    return (
-        "BIN-004",
-        Severity.LOW,
-        "Compiled Python bytecode artifact present",
-        (
-            "Prefer scanning source files with bytecode provenance; "
-            "review for hidden or mismatched source content."
-        ),
-    )
-
-
-def _binary_artifact_findings(artifacts: list[BinaryArtifact]) -> list[Finding]:
-    grouped: dict[str, list[BinaryArtifact]] = {}
-    for item in artifacts:
-        grouped.setdefault(item.kind, []).append(item)
-    findings: list[Finding] = []
-    for kind, items in grouped.items():
-        rule_id, severity, title, mitigation = _binary_kind_template(kind)
-        sample = ", ".join(str(item.path) for item in items[:3])
-        findings.append(
-            Finding(
-                id=rule_id,
-                category="binary_artifact",
-                severity=severity,
-                confidence=0.9,
-                title=title,
-                evidence_path=str(items[0].path),
-                snippet=f"count={len(items)} sample={sample}",
-                mitigation=mitigation,
-            )
-        )
-    return findings
-
-
-def _ip_in_cidrs(ip_value: str, cidrs: list[str]) -> bool:
-    try:
-        ip_obj = ipaddress.ip_address(ip_value)
-    except ValueError:
-        return False
-    for cidr in cidrs:
-        try:
-            if ip_obj in ipaddress.ip_network(cidr, strict=False):
-                return True
-        except ValueError:
-            continue
-    return False
-
-
-def scan(
-    target: Path | str,
-    policy: Policy,
-    policy_source: str,
-    *,
-    url_max_links: int = 25,
-    url_timeout_seconds: int = 12,
-    url_same_origin_only: bool = True,
-    clamav: bool = False,
-    clamav_timeout_seconds: int = 30,
-    ml_detect: bool = False,
-    rulepack_channel: str = "stable",
-    graph_scan: bool = False,
-    max_file_size_bytes: int = 1_048_576,  # 1 MB default — skip larger files with a warning
-    file_timeout_seconds: int = 30,  # per-file rule-matching timeout
-) -> ScanReport:
-    prepared = prepare_target(
-        target,
-        policy,
-        url_max_links=url_max_links,
-        url_timeout_seconds=url_timeout_seconds,
-        url_same_origin_only=url_same_origin_only,
-    )
-    try:
-        ruleset: CompiledRulePack = load_compiled_builtin_rulepack(channel=rulepack_channel)
-        inventory = iter_text_files(
-            prepared.root,
-            policy.limits["max_files"],
-            policy.limits["max_bytes"],
-            policy.limits.get("max_binary_artifacts", 500),
-            policy.limits.get("max_binary_bytes", 100_000_000),
-        )
-        files = inventory.text_files
-        findings: list[Finding] = []
-        iocs: list[IOC] = []
-        capabilities: list[Capability] = []
-        dependency_findings: list[DependencyFinding] = []
-        # Triage score accumulators — track max across all files
-        _triage_sem_inj: float = 0.0
-        _triage_se: float = 0.0
-        _triage_ml_prob: float | None = None
-        vuln_db = _load_builtin_vuln_db()
-        ioc_db = _load_builtin_ioc_db()
-        ioc_db, vuln_db, intel_sources = _merge_user_intel(ioc_db, vuln_db)
-        findings.extend(_binary_artifact_findings(inventory.binary_artifacts))
-
-        if clamav:
-            clamav_result = clamav_scan_paths(prepared.root, timeout_seconds=clamav_timeout_seconds)
-            if not clamav_result.available:
-                findings.append(
-                    Finding(
-                        id="AV-UNAVAILABLE",
-                        category="artifact_malware",
-                        severity=Severity.LOW,
-                        confidence=1.0,
-                        title="ClamAV unavailable",
-                        evidence_path=str(prepared.root),
-                        snippet=(clamav_result.message or "ClamAV unavailable")[:220],
-                        mitigation="Install ClamAV (`clamscan`) or disable --clamav for this run.",
-                    )
-                )
-            else:
-                for detection in clamav_result.detections:
-                    findings.append(
-                        Finding(
-                            id="AV-001",
-                            category="artifact_malware",
-                            severity=Severity.HIGH,
-                            confidence=0.9,
-                            title="ClamAV malware detection",
-                            evidence_path=detection.path,
-                            snippet=detection.signature[:220],
-                            mitigation=(
-                                "Treat detected artifact as malicious and remove/quarantine before execution."
-                            ),
-                        )
-                    )
-
-        # Advisory: if script files are present but ClamAV was not requested,
-        # emit a single LOW advisory so the report can recommend a deeper scan.
-        if not clamav:
-            script_files = [p for p in files if p.suffix.lower() in SCRIPT_SUFFIXES]
-            if script_files:
-                sample = ", ".join(p.name for p in script_files[:3])
-                if len(script_files) > 3:
-                    sample += f" (+{len(script_files) - 3} more)"
-                findings.append(
-                    Finding(
-                        id="AV-ADVISORY",
-                        category="artifact_malware",
-                        severity=Severity.LOW,
-                        confidence=1.0,
-                        title="Bundled script files detected — ClamAV scan not performed",
-                        evidence_path=str(prepared.root),
-                        snippet=sample,
-                        mitigation=(
-                            "This skill bundles executable script files. "
-                            "Re-run with --clamav to perform a malware artifact scan on these files. "
-                            "Alternatively, manually inspect the scripts before loading this skill."
-                        ),
-                    )
-                )
-
-        def _scan_one_file(
-            path: Path,
-        ) -> tuple[
-            list[Finding],
-            list[IOC],
-            list[Capability],
-            list[DependencyFinding],
-            float,  # sem_inj triage score
-            float,  # se triage score
-            float | None,  # ml prob
-        ]:
-            """Scan a single file and return all findings/IOCs/capabilities.
-
-            Designed to run inside a ThreadPoolExecutor so the caller can
-            enforce a per-file wall-clock timeout via Future.result(timeout=...).
-            """
-            _f: list[Finding] = []
-            _iocs: list[IOC] = []
-            _caps: list[Capability] = []
-            _deps: list[DependencyFinding] = []
-            _sem: float = 0.0
-            _se: float = 0.0
-            _ml: float | None = None
-
-            # Per-file size guard
-            try:
-                file_bytes = path.stat().st_size
-            except OSError:
-                file_bytes = 0
-            if file_bytes > max_file_size_bytes:
-                _f.append(
-                    Finding(
-                        id="SCAN-SIZE-SKIP",
-                        category="scanner_advisory",
-                        severity=Severity.LOW,
-                        confidence=1.0,
-                        title="File skipped — exceeds max file size",
-                        evidence_path=str(path),
-                        snippet=(
-                            f"{file_bytes // 1024} KB > "
-                            f"{max_file_size_bytes // 1024} KB limit. "
-                            "Re-run with --max-file-size to raise the limit."
-                        ),
-                        mitigation="Inspect the file manually or raise --max-file-size.",
-                    )
-                )
-                return _f, _iocs, _caps, _deps, _sem, _se, _ml
-
-            text = _safe_read_text(path)
-            if not text:
-                return _f, _iocs, _caps, _deps, _sem, _se, _ml
-            analysis_text = _prepare_analysis_text(text)
-
-            file_ext = path.suffix.lower()
-            for rule in ruleset.static_rules:
-                if rule.graph_rule:
-                    continue
-                if rule.language is not None:
-                    allowed_exts = _LANG_EXTENSIONS.get(rule.language)
-                    if allowed_exts is not None and file_ext not in allowed_exts:
-                        continue
-                if rule.multiline:
-                    # Full-text match for patterns that span multiple lines
-                    m = rule.pattern.search(analysis_text)
-                    if m:
-                        matched_text = m.group(0)
-                        line_no = analysis_text[: m.start()].count("\n") + 1
-                        _f.append(
-                            Finding(
-                                id=rule.id,
-                                category=rule.category,
-                                severity=rule.severity,
-                                confidence=rule.confidence,
-                                title=rule.title,
-                                evidence_path=str(path),
-                                line=line_no,
-                                snippet=matched_text.replace("\n", " ").strip()[:240],
-                                mitigation=rule.mitigation,
-                            )
-                        )
-                else:
-                    for line_no, line in enumerate(analysis_text.splitlines(), 1):
-                        if rule.pattern.search(line):
-                            _f.append(
-                                Finding(
-                                    id=rule.id,
-                                    category=rule.category,
-                                    severity=rule.severity,
-                                    confidence=rule.confidence,
-                                    title=rule.title,
-                                    evidence_path=str(path),
-                                    line=line_no,
-                                    snippet=line.strip()[:240],
-                                    mitigation=rule.mitigation,
-                                )
-                            )
-                            break
-
-            _f.extend(local_prompt_injection_findings(path, analysis_text))
-            _f.extend(local_social_engineering_findings(path, analysis_text))
-            _sem = classify_prompt_injection_raw(analysis_text)
-            _se = classify_social_engineering_raw(analysis_text)
-
-            if ml_detect:
-                ml_findings = ml_prompt_injection_findings(path, analysis_text)
-                _f.extend(ml_findings)
-                for mf in ml_findings:
-                    if mf.id.startswith("ML-"):
-                        _ml = max(_ml or 0.0, mf.confidence)
-
-            for capability_name, pattern in ruleset.capability_patterns:
-                if pattern.search(analysis_text):
-                    _caps.append(
-                        Capability(name=capability_name, evidence_path=str(path), detail="Pattern match")
-                    )
-
-            if file_ext == ".py":
-                for ast_finding in detect_python_ast_flows(analysis_text):
-                    _f.append(
-                        Finding(
-                            id=ast_finding.id,
-                            category=ast_finding.category,
-                            severity=ast_finding.severity,
-                            confidence=ast_finding.confidence,
-                            title=ast_finding.title,
-                            evidence_path=str(path),
-                            line=ast_finding.line,
-                            snippet=ast_finding.snippet,
-                            mitigation=ast_finding.mitigation,
-                        )
-                    )
-
-            _iocs.extend(_extract_iocs(path, analysis_text))
-
-            # Chain rules with windowed proximity constraint
-            _window_cache: dict[int, list[set[str]]] = {}
-
-            def _get_windows(wl: int) -> list[set[str]]:
-                if wl not in _window_cache:
-                    _window_cache[wl] = _extract_actions_windowed(
-                        analysis_text, ruleset.action_patterns, window_lines=wl
-                    )
-                return _window_cache[wl]
-
-            fired_chain_ids: set[str] = set()
-            for chain_rule in ruleset.chain_rules:
-                if chain_rule.id in fired_chain_ids:
-                    continue
-                effective_window = (
-                    chain_rule.window_lines if chain_rule.window_lines is not None else _CHAIN_WINDOW_LINES
-                )
-                for window_actions in _get_windows(effective_window):
-                    if chain_rule.all_of.issubset(window_actions):
-                        fired_chain_ids.add(chain_rule.id)
-                        _f.append(
-                            Finding(
-                                id=chain_rule.id,
-                                category=chain_rule.category,
-                                severity=chain_rule.severity,
-                                confidence=chain_rule.confidence,
-                                title=chain_rule.title,
-                                evidence_path=str(path),
-                                snippet=chain_rule.snippet or " + ".join(sorted(chain_rule.all_of)),
-                                mitigation=chain_rule.mitigation,
-                                chain_actions=sorted(chain_rule.all_of),
-                            )
-                        )
-                        break
-
-            lower = path.name.lower()
-            if lower == "requirements.txt":
-                for name, version in _parse_requirements(text):
-                    if name in vuln_db.get("python", {}) and version in vuln_db["python"][name]:
-                        entry = vuln_db["python"][name][version]
-                        _deps.append(
-                            DependencyFinding(
-                                ecosystem="python",
-                                name=name,
-                                version=version,
-                                vulnerability_id=entry["id"],
-                                severity=Severity(entry["severity"]),
-                                fixed_version=entry.get("fixed"),
-                            )
-                        )
-                for name, spec in _find_unpinned_requirements(text):
-                    _f.append(
-                        Finding(
-                            id="DEP-UNPIN",
-                            category="dependency_vulnerability",
-                            severity=Severity.MEDIUM,
-                            confidence=0.8,
-                            title=f"Unpinned python dependency: {name}",
-                            evidence_path=str(path),
-                            snippet=spec,
-                            mitigation=(
-                                "Pin exact dependency versions to improve reproducibility "
-                                "and reduce supply-chain risk."
-                            ),
-                        )
-                    )
-            if lower == "package.json":
-                for name, version in _parse_package_json(text):
-                    version_norm = version.lstrip("^~>=< ")
-                    if name in vuln_db.get("npm", {}) and version_norm in vuln_db["npm"][name]:
-                        entry = vuln_db["npm"][name][version_norm]
-                        _deps.append(
-                            DependencyFinding(
-                                ecosystem="npm",
-                                name=name,
-                                version=version,
-                                vulnerability_id=entry["id"],
-                                severity=Severity(entry["severity"]),
-                                fixed_version=entry.get("fixed"),
-                            )
-                        )
-                    if _is_unpinned_npm(version):
-                        _f.append(
-                            Finding(
-                                id="DEP-UNPIN",
-                                category="dependency_vulnerability",
-                                severity=Severity.MEDIUM,
-                                confidence=0.75,
-                                title=f"Unpinned npm dependency: {name}",
-                                evidence_path=str(path),
-                                snippet=f"{name}: {version}",
-                                mitigation=(
-                                    "Pin exact dependency versions to improve reproducibility "
-                                    "and reduce supply-chain risk."
-                                ),
-                            )
-                        )
-                scripts = _parse_package_scripts(text)
-                lifecycle_hooks = ("preinstall", "install", "postinstall", "prepare")
-                for hook in lifecycle_hooks:
-                    cmd = scripts.get(hook)
-                    if not cmd:
-                        continue
-                    if RISKY_NPM_SCRIPT_RE.search(cmd):
-                        _f.append(
-                            Finding(
-                                id="SUP-001",
-                                category="malware_pattern",
-                                severity=Severity.HIGH,
-                                confidence=0.88,
-                                title=f"Risky npm lifecycle script: {hook}",
-                                evidence_path=str(path),
-                                snippet=f"{hook}: {cmd[:220]}",
-                                mitigation=(
-                                    "Remove network/bootstrap execution from lifecycle hooks. "
-                                    "Use explicit reviewed setup commands instead."
-                                ),
-                            )
-                        )
-
-            return _f, _iocs, _caps, _deps, _sem, _se, _ml
-
-        # Dispatch per-file scan with individual timeout using a thread pool.
-        # Each file gets file_timeout_seconds; if it exceeds the limit a
-        # SCAN-TIMEOUT-SKIP advisory finding is emitted and the file is skipped.
-        with concurrent.futures.ThreadPoolExecutor(max_workers=1) as _executor:
-            _futures = {_executor.submit(_scan_one_file, p): p for p in files}
-            for _fut, _path in _futures.items():
-                try:
-                    _ff, _fi, _fc, _fd, _sem, _se, _ml = _fut.result(timeout=file_timeout_seconds)
-                    findings.extend(_ff)
-                    iocs.extend(_fi)
-                    capabilities.extend(_fc)
-                    dependency_findings.extend(_fd)
-                    _triage_sem_inj = max(_triage_sem_inj, _sem)
-                    _triage_se = max(_triage_se, _se)
-                    if _ml is not None:
-                        _triage_ml_prob = max(_triage_ml_prob or 0.0, _ml)
-                except concurrent.futures.TimeoutError:
-                    findings.append(
-                        Finding(
-                            id="SCAN-TIMEOUT-SKIP",
-                            category="scanner_advisory",
-                            severity=Severity.LOW,
-                            confidence=1.0,
-                            title="File skipped — per-file scan timeout exceeded",
-                            evidence_path=str(_path),
-                            snippet=(
-                                f"Scan exceeded {file_timeout_seconds}s limit. "
-                                "Re-run with --timeout to raise the limit."
-                            ),
-                            mitigation="Inspect the file manually or raise --timeout.",
-                        )
-                    )
-
-        dedup_iocs: dict[tuple[str, str, str], IOC] = {}
-        allow_domains = {d.lower() for d in policy.allow_domains}
-        for ioc in iocs:
-            key = (ioc.kind, ioc.value.lower(), ioc.source_path)
-            listed = False
-            if ioc.kind == "domain" and ioc.value.lower() in ioc_db.get("domains", []):
-                listed = True
-            if ioc.kind == "ip" and ioc.value in ioc_db.get("ips", []):
-                listed = True
-            if ioc.kind == "ip" and _ip_in_cidrs(ioc.value, ioc_db.get("cidrs", [])):
-                listed = True
-            if ioc.kind == "url" and ioc.value.lower() in ioc_db.get("urls", []):
-                listed = True
-            if ioc.kind == "domain" and ioc.value.lower() in allow_domains:
-                listed = False
-            ioc.listed = listed
-            dedup_iocs[key] = ioc
-        iocs = list(dedup_iocs.values())
-
-        block_domains = {d.lower() for d in policy.block_domains}
-        for ioc in iocs:
-            if ioc.kind == "domain" and ioc.value.lower() in block_domains:
-                findings.append(
-                    Finding(
-                        id="POL-IOC-BLOCK",
-                        category="threat_intel",
-                        severity=Severity.HIGH,
-                        confidence=1.0,
-                        title="Domain blocked by local policy",
-                        evidence_path=ioc.source_path,
-                        snippet=ioc.value,
-                        mitigation=(
-                            "Replace blocked destination with an approved domain "
-                            "or remove network dependency."
-                        ),
-                    )
-                )
-
-        for ioc in iocs:
-            if ioc.listed:
-                findings.append(
-                    Finding(
-                        id="IOC-001",
-                        category="threat_intel",
-                        severity=Severity.HIGH,
-                        confidence=0.95,
-                        title="IOC matched local blocklist",
-                        evidence_path=ioc.source_path,
-                        snippet=ioc.value,
-                        mitigation=(
-                            "Block install/use and investigate indicator reputation. "
-                            "Remove all references to this IOC."
-                        ),
-                    )
-                )
-
-        for dep in dependency_findings:
-            findings.append(
-                Finding(
-                    id="DEP-001",
-                    category="dependency_vulnerability",
-                    severity=dep.severity,
-                    confidence=0.9,
-                    title=f"Vulnerable dependency: {dep.name}@{dep.version}",
-                    evidence_path=dep.name,
-                    snippet=dep.vulnerability_id,
-                    mitigation="Upgrade to a non-vulnerable dependency version and refresh lockfiles.",
-                )
-            )
-
-        if prepared.read_warnings:
-            findings.append(
-                Finding(
-                    id="SRC-READ-ERR",
-                    category="source_access",
-                    severity=Severity.LOW,
-                    confidence=1.0,
-                    title="Some linked sources could not be fetched",
-                    evidence_path=str(target),
-                    snippet=f"unreadable_links={len(prepared.read_warnings)}",
-                    mitigation=(
-                        "Verify linked source URLs are reachable and public. "
-                        "Unreachable links are not treated as malicious by default."
-                    ),
-                )
-            )
-
-        if prepared.policy_warnings:
-            for warning in prepared.policy_warnings:
-                if warning.startswith("BIN-OPAQUE-002:"):
-                    archive_name = warning.split(":", 2)[1]
-                    findings.append(
-                        Finding(
-                            id="BIN-OPAQUE-002",
-                            category="binary_artifact",
-                            severity=Severity.MEDIUM,
-                            confidence=0.9,
-                            title="Password-protected archive — contents unverified",
-                            evidence_path=str(target),
-                            snippet=archive_name,
-                            mitigation=(
-                                "The archive is password-protected and could not be extracted for scanning. "
-                                "Manually inspect the contents before loading this skill."
-                            ),
-                        )
-                    )
-                elif warning.startswith("BIN-OPAQUE-001:"):
-                    parts = warning.split(":", 3)
-                    archive_name = parts[1] if len(parts) > 1 else str(target)
-                    reason = parts[2] if len(parts) > 2 else "unsupported format"
-                    findings.append(
-                        Finding(
-                            id="BIN-OPAQUE-001",
-                            category="binary_artifact",
-                            severity=Severity.MEDIUM,
-                            confidence=0.8,
-                            title="Archive format not extractable — contents unverified",
-                            evidence_path=str(target),
-                            snippet=f"{archive_name}: {reason}",
-                            mitigation=(
-                                "Install skillscan[archives] for expanded archive support, "
-                                "or manually inspect the archive contents before loading this skill."
-                            ),
-                        )
-                    )
-                else:
-                    # URL skip policy warning (legacy path)
-                    findings.append(
-                        Finding(
-                            id="URL-SKIP-POLICY",
-                            category="source_access",
-                            severity=Severity.LOW,
-                            confidence=1.0,
-                            title="Some linked sources were skipped by URL safety policy",
-                            evidence_path=str(target),
-                            snippet=f"skipped_links={len(prepared.policy_warnings)}",
-                            mitigation=(
-                                "Links skipped due to same-origin policy. "
-                                "Use --url-same-origin-only false to include cross-origin links when needed."
-                            ),
-                        )
-                    )
-                    break  # one finding per batch of URL warnings
-
-        if graph_scan:
-            from skillscan.detectors.skill_graph import skill_graph_findings
-
-            findings.extend(skill_graph_findings(prepared.root))
-
-        score = 0
-        block_score = 0
-        for finding in findings:
-            weight = policy.weights.get(finding.category, 1)
-            contribution = SEVERITY_SCORE[finding.severity] * weight
-            score += contribution
-            if finding.confidence >= policy.block_min_confidence:
-                block_score += contribution
-
-        if any(
-            f.id in policy.hard_block_rules and f.confidence >= policy.block_min_confidence for f in findings
-        ):
-            verdict = Verdict.BLOCK
-        elif block_score >= policy.thresholds["block"]:
-            verdict = Verdict.BLOCK
-        elif score >= policy.thresholds["warn"]:
-            verdict = Verdict.WARN
-        else:
-            verdict = Verdict.ALLOW
-
-        metadata = ScanMetadata(
-            scanner_version=__version__,
-            target=str(target),
-            target_type=prepared.target_type,
-            ecosystem_hints=detect_ecosystems(prepared.root),
-            rulepack_version=ruleset.version,
-            policy_profile=policy.name,
-            policy_source=policy_source,
-            intel_sources=intel_sources,
-        )
-
-        # Build triage metadata — compute has_sub_threshold_signal
-        _has_sub = (
-            (_triage_sem_inj > 0.25)
-            or (_triage_se > 0.25)
-            or (_triage_ml_prob is not None and _triage_ml_prob > 0.30)
-        )
-        triage_metadata = TriageMetadata(
-            semantic_injection_score=round(_triage_sem_inj, 4),
-            social_engineering_score=round(_triage_se, 4),
-            ml_injection_probability=(round(_triage_ml_prob, 4) if _triage_ml_prob is not None else None),
-            has_sub_threshold_signal=_has_sub,
-        )
-
-        return ScanReport(
-            metadata=metadata,
-            verdict=verdict,
-            score=score,
-            findings=findings,
-            iocs=iocs,
-            dependency_findings=dependency_findings,
-            capabilities=capabilities,
-            triage_metadata=triage_metadata,
-        )
-    finally:
-        if prepared.cleanup_dir is not None:
-            prepared.cleanup_dir.cleanup()
-        os.environ.pop("PYTHONINSPECT", None)
+__all__ = [
+    "ArchiveFormatError",
+    "ArchivePasswordError",
+    "BinaryArtifact",
+    "FileInventory",
+    "PreparedTarget",
+    "ScanError",
+    "_decode_base64_fragments",
+    "_extract_actions",
+    "_extract_actions_windowed",
+    "_extract_iocs",
+    "_find_unpinned_requirements",
+    "_ip_in_cidrs",
+    "_merge_user_intel",
+    "_normalize_text",
+    "_prepare_analysis_text",
+    "_safe_extract_7z",
+    "_safe_extract_tar",
+    "_safe_extract_zip",
+    "_safe_extract_zst",
+    "iter_text_files",
+    "prepare_target",
+    "clamav_scan_paths",
+    "scan",
+]
