+++ src/skillscan/analysis_pkg/__init__.py
@@ -1,0 +1,37 @@
+# analysis/__init__.py
+# Compatibility shim: re-exports all public symbols so existing imports work unchanged.
+from skillscan.analysis_pkg._archive import (
+    ArchiveFormatError,
+    ArchivePasswordError,
+    BinaryArtifact,
+    FileInventory,
+    PreparedTarget,
+    ScanError,
+    prepare_target,
+)
+from skillscan.analysis_pkg._scanner import scan
+from skillscan.analysis_pkg._text import (
+    _extract_actions_windowed,
+    _extract_iocs,
+    _find_unpinned_requirements,
+    _ip_in_cidrs,
+    _merge_user_intel,
+    iter_text_files,
+)
+
+__all__ = [
+    "ArchiveFormatError",
+    "ArchivePasswordError",
+    "BinaryArtifact",
+    "FileInventory",
+    "PreparedTarget",
+    "ScanError",
+    "prepare_target",
+    "_extract_actions_windowed",
+    "_extract_iocs",
+    "_find_unpinned_requirements",
+    "_ip_in_cidrs",
+    "_merge_user_intel",
+    "iter_text_files",
+    "scan",
+]
