+++ src/skillscan/analysis_pkg/_archive.py
@@ -1,0 +1,323 @@
+# analysis/_archive.py — archive extraction helpers, dataclasses, and prepare_target
+from __future__ import annotations
+
+import tarfile
+import tempfile
+import zipfile
+from dataclasses import dataclass
+from pathlib import Path
+
+from skillscan.models import (
+    Policy,
+    detect_archive_format,
+    is_archive,
+)
+from skillscan.remote import RemoteFetchError, fetch_remote_target, is_url_target
+
+# Binary-classification suffix sets (also imported by _text.py)
+BYTECODE_SUFFIXES = {".pyc", ".pyo"}
+SCRIPT_SUFFIXES = {
+    ".py",
+    ".sh",
+    ".bash",
+    ".rb",
+    ".js",
+    ".ts",
+    ".mjs",
+    ".cjs",
+    ".go",
+    ".rs",
+    ".pl",
+    ".ps1",
+    ".psm1",
+}
+LIBRARY_SUFFIXES = {".so", ".dll", ".dylib", ".a"}
+EXECUTABLE_SUFFIXES = {".exe", ".msi", ".com"}
+KNOWN_BINARY_SUFFIXES = {
+    ".bin",
+    ".o",
+    ".obj",
+    ".class",
+    ".jar",
+    ".wasm",
+    ".pyz",
+    ".whl",
+}
+MACHO_MAGICS = (
+    b"\xfe\xed\xfa\xce",
+    b"\xce\xfa\xed\xfe",
+    b"\xfe\xed\xfa\xcf",
+    b"\xcf\xfa\xed\xfe",
+    b"\xca\xfe\xba\xbe",
+    b"\xbe\xba\xfe\xca",
+)
+
+
+@dataclass
+class PreparedTarget:
+    root: Path
+    target_type: str
+    cleanup_dir: tempfile.TemporaryDirectory[str] | None
+    read_warnings: list[str]
+    policy_warnings: list[str]
+
+
+@dataclass
+class BinaryArtifact:
+    path: Path
+    kind: str
+    detail: str
+
+
+@dataclass
+class FileInventory:
+    text_files: list[Path]
+    binary_artifacts: list[BinaryArtifact]
+
+
+class ScanError(Exception):
+    pass
+
+
+class ArchivePasswordError(Exception):
+    """Raised when an archive is password-protected and cannot be extracted."""
+
+
+class ArchiveFormatError(Exception):
+    """Raised when an archive format is not supported by available libraries."""
+
+
+def _safe_extract_zip(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
+    total = 0
+    with zipfile.ZipFile(src) as zf:
+        infos = zf.infolist()
+        if len(infos) > max_files:
+            raise ScanError(f"Archive has too many files: {len(infos)}")
+        for info in infos:
+            name = info.filename
+            if name.startswith("/") or ".." in Path(name).parts:
+                raise ScanError(f"Unsafe archive path: {name}")
+            total += info.file_size
+            if total > max_bytes:
+                raise ScanError("Archive exceeds max bytes limit")
+            zf.extract(info, dst)
+
+
+def _safe_extract_tar(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
+    total = 0
+    with tarfile.open(src) as tf:
+        members = tf.getmembers()
+        if len(members) > max_files:
+            raise ScanError(f"Archive has too many files: {len(members)}")
+        for member in members:
+            p = Path(member.name)
+            if p.is_absolute() or ".." in p.parts:
+                raise ScanError(f"Unsafe archive path: {member.name}")
+            if member.issym() or member.islnk():
+                raise ScanError(f"Symlink/hardlink not allowed in archive: {member.name}")
+            total += member.size
+            if total > max_bytes:
+                raise ScanError("Archive exceeds max bytes limit")
+        # filter='data' was added in Python 3.12; use keyword argument only when available
+        if hasattr(tf, "extraction_filter"):
+            tf.extractall(dst, filter="data")
+        else:
+            for member in members:
+                tf.extract(member, dst)
+
+
+def _safe_extract_7z(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
+    try:
+        import py7zr  # optional extra
+    except ImportError as exc:
+        raise ArchiveFormatError("py7zr not installed; install skillscan[archives]") from exc
+    try:
+        with py7zr.SevenZipFile(src, mode="r") as sz:
+            entries = sz.list()
+            if len(entries) > max_files:
+                raise ScanError(f"Archive has too many files: {len(entries)}")
+            total = sum(e.uncompressed for e in entries if e.uncompressed)
+            if total > max_bytes:
+                raise ScanError("Archive exceeds max bytes limit")
+            for e in entries:
+                p = Path(e.filename)
+                if p.is_absolute() or ".." in p.parts:
+                    raise ScanError(f"Unsafe archive path: {e.filename}")
+            sz.extractall(path=dst)
+    except py7zr.exceptions.PasswordRequired as exc:
+        raise ArchivePasswordError("7z archive is password-protected") from exc
+
+
+def _safe_extract_rar(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
+    try:
+        import rarfile  # optional extra
+    except ImportError as exc:
+        raise ArchiveFormatError("rarfile not installed; install skillscan[archives]") from exc
+    try:
+        with rarfile.RarFile(src) as rf:
+            infos = rf.infolist()
+            if len(infos) > max_files:
+                raise ScanError(f"Archive has too many files: {len(infos)}")
+            total = sum(i.file_size for i in infos)
+            if total > max_bytes:
+                raise ScanError("Archive exceeds max bytes limit")
+            for info in infos:
+                p = Path(info.filename)
+                if p.is_absolute() or ".." in p.parts:
+                    raise ScanError(f"Unsafe archive path: {info.filename}")
+            rf.extractall(path=dst)
+    except rarfile.PasswordRequired as exc:
+        raise ArchivePasswordError("RAR archive is password-protected") from exc
+    except rarfile.NeedFirstVolume as exc:
+        raise ArchiveFormatError("Multi-volume RAR not supported") from exc
+
+
+def _safe_extract_xz(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
+    """Extract .xz (plain xz-compressed file, not tar.xz — that's handled by tarfile)."""
+    import lzma
+
+    out_name = src.stem  # strip .xz
+    out_path = dst / out_name
+    total = 0
+    with lzma.open(src, "rb") as fh:
+        while True:
+            chunk = fh.read(65536)
+            if not chunk:
+                break
+            total += len(chunk)
+            if total > max_bytes:
+                raise ScanError("Archive exceeds max bytes limit")
+            with out_path.open("ab") as out:
+                out.write(chunk)
+
+
+def _safe_extract_bz2(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
+    """Extract .bz2 (plain bzip2-compressed file, not tar.bz2 — that's handled by tarfile)."""
+    import bz2
+
+    out_name = src.stem
+    out_path = dst / out_name
+    total = 0
+    with bz2.open(src, "rb") as fh:
+        while True:
+            chunk = fh.read(65536)
+            if not chunk:
+                break
+            total += len(chunk)
+            if total > max_bytes:
+                raise ScanError("Archive exceeds max bytes limit")
+            with out_path.open("ab") as out:
+                out.write(chunk)
+
+
+def _safe_extract_zst(src: Path, dst: Path, max_files: int, max_bytes: int) -> None:
+    """Extract .zst (plain Zstandard-compressed file)."""
+    try:
+        import zstandard  # optional extra
+    except ImportError as exc:
+        raise ArchiveFormatError("zstandard not installed; install skillscan[archives]") from exc
+    out_name = src.stem
+    out_path = dst / out_name
+    total = 0
+    dctx = zstandard.ZstdDecompressor()
+    with src.open("rb") as ifh, out_path.open("wb") as ofh:
+        for chunk in dctx.read_to_iter(ifh, read_size=65536):
+            total += len(chunk)
+            if total > max_bytes:
+                raise ScanError("Archive exceeds max bytes limit")
+            ofh.write(chunk)
+
+
+def prepare_target(
+    target: Path | str,
+    policy: Policy,
+    url_max_links: int = 25,
+    url_timeout_seconds: int = 12,
+    url_same_origin_only: bool = True,
+) -> PreparedTarget:
+    if isinstance(target, str) and is_url_target(target):
+        try:
+            fetched = fetch_remote_target(
+                target,
+                max_links=url_max_links,
+                timeout_seconds=url_timeout_seconds,
+                same_origin_only=url_same_origin_only,
+            )
+        except RemoteFetchError as exc:
+            raise ScanError(str(exc)) from exc
+        return PreparedTarget(
+            root=fetched.root,
+            target_type="url",
+            cleanup_dir=fetched.cleanup_dir,
+            read_warnings=fetched.unreadable_urls,
+            policy_warnings=fetched.skipped_urls,
+        )
+
+    if isinstance(target, str):
+        target = Path(target)
+
+    if not target.exists():
+        raise ScanError(f"Target does not exist: {target}")
+    if target.is_dir():
+        return PreparedTarget(
+            root=target,
+            target_type="directory",
+            cleanup_dir=None,
+            read_warnings=[],
+            policy_warnings=[],
+        )
+    if target.is_file() and not is_archive(target):
+        tmp = tempfile.TemporaryDirectory(prefix="skillscan-")
+        dst = Path(tmp.name)
+        (dst / target.name).write_bytes(target.read_bytes())
+        return PreparedTarget(
+            root=dst,
+            target_type="file",
+            cleanup_dir=tmp,
+            read_warnings=[],
+            policy_warnings=[],
+        )
+    if target.is_file() and is_archive(target):
+        tmp = tempfile.TemporaryDirectory(prefix="skillscan-")
+        dst = Path(tmp.name)
+        fmt = detect_archive_format(target)
+        policy_warnings: list[str] = []
+        try:
+            if fmt == "zip":
+                _safe_extract_zip(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
+            elif fmt in ("gz", "bz2", "xz", "tar"):
+                # tarfile handles .tar, .tar.gz, .tar.bz2, .tar.xz transparently
+                try:
+                    _safe_extract_tar(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
+                except tarfile.ReadError:
+                    # Plain compressed file (not tar) — fall through to single-file decompressors
+                    if fmt == "xz":
+                        _safe_extract_xz(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
+                    elif fmt == "bz2":
+                        _safe_extract_bz2(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
+                    else:
+                        raise
+            elif fmt == "7z":
+                _safe_extract_7z(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
+            elif fmt == "rar":
+                _safe_extract_rar(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
+            elif fmt == "zst":
+                _safe_extract_zst(target, dst, policy.limits["max_files"], policy.limits["max_bytes"])
+            else:
+                raise ArchiveFormatError(f"Unsupported archive format: {fmt}")
+        except ArchivePasswordError:
+            # Password-protected: copy as-is and emit BIN-OPAQUE-002 during scan
+            (dst / target.name).write_bytes(target.read_bytes())
+            policy_warnings.append(f"BIN-OPAQUE-002:{target.name}")
+        except ArchiveFormatError as exc:
+            # Library not available or format not supported: copy as-is and emit BIN-OPAQUE-001
+            (dst / target.name).write_bytes(target.read_bytes())
+            policy_warnings.append(f"BIN-OPAQUE-001:{target.name}:{exc}")
+        return PreparedTarget(
+            root=dst,
+            target_type="archive",
+            cleanup_dir=tmp,
+            read_warnings=[],
+            policy_warnings=policy_warnings,
+        )
+    raise ScanError("Unsupported target type")
