+++ src/skillscan/analysis_pkg/_text.py
@@ -1,0 +1,476 @@
+# analysis/_text.py — module-level constants, text processing, IOC/dep extraction, intel helpers
+from __future__ import annotations
+
+import base64
+import binascii
+import ipaddress
+import json
+import re
+import unicodedata
+from importlib import resources
+from pathlib import Path
+from typing import cast
+
+from skillscan.analysis_pkg._archive import (
+    BYTECODE_SUFFIXES,
+    EXECUTABLE_SUFFIXES,
+    KNOWN_BINARY_SUFFIXES,
+    LIBRARY_SUFFIXES,
+    MACHO_MAGICS,
+    BinaryArtifact,
+    FileInventory,
+    ScanError,
+)
+from skillscan.intel import load_store
+from skillscan.models import (
+    IOC,
+    Finding,
+    Severity,
+)
+
+URL_RE = re.compile(r"https?://[^\s\"'<>]+", re.IGNORECASE)
+IP_RE = re.compile(r"\b(?:\d{1,3}\.){3}\d{1,3}\b")
+DOMAIN_RE = re.compile(r"\b[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}\b")
+ZERO_WIDTH_RE = re.compile(r"[\u200b-\u200f\u2060\ufeff]")
+B64_TOKEN_RE = re.compile(r"(?<![A-Za-z0-9+/_-])[A-Za-z0-9+/_-]{16,}={0,2}(?![A-Za-z0-9+/_=-])")
+QUOTED_B64_FRAGMENT_RE = re.compile(r"[\"']([A-Za-z0-9+/_-]{8,}={0,2})[\"']")
+RISKY_NPM_SCRIPT_RE = re.compile(
+    r"\b(curl|wget|invoke-webrequest|powershell|pwsh|bash|sh|cmd\.exe|certutil|bitsadmin|nc|ncat|netcat)\b|"
+    r"https?://",
+    re.IGNORECASE,
+)
+KNOWN_TLDS = {
+    "ai",
+    "app",
+    "biz",
+    "co",
+    "com",
+    "dev",
+    "edu",
+    "fr",
+    "gov",
+    "info",
+    "io",
+    "jp",
+    "net",
+    "org",
+    "ru",
+    "uk",
+    "us",
+    "xyz",
+}
+
+SEVERITY_SCORE = {
+    Severity.LOW: 5,
+    Severity.MEDIUM: 15,
+    Severity.HIGH: 35,
+    Severity.CRITICAL: 60,
+}
+
+IOCDB = dict[str, list[str]]
+VulnRecord = dict[str, str]
+VulnVersionMap = dict[str, VulnRecord]
+VulnPackageMap = dict[str, VulnVersionMap]
+VulnDB = dict[str, VulnPackageMap]
+
+
+def _read_head(path: Path, limit: int = 4096) -> bytes:
+    try:
+        with path.open("rb") as fh:
+            return fh.read(limit)
+    except OSError:
+        return b""
+
+
+def _classify_non_text(path: Path) -> BinaryArtifact | None:
+    suffix = path.suffix.lower()
+    if "__pycache__" in path.parts or suffix in BYTECODE_SUFFIXES:
+        return BinaryArtifact(path=path, kind="python_bytecode", detail="compiled python bytecode artifact")
+    if suffix in LIBRARY_SUFFIXES:
+        return BinaryArtifact(path=path, kind="binary_library", detail="compiled library artifact")
+    if suffix in EXECUTABLE_SUFFIXES:
+        return BinaryArtifact(path=path, kind="executable_binary", detail="known executable extension")
+
+    head = _read_head(path)
+    if not head:
+        return None
+    if head.startswith(b"\x7fELF") or head.startswith(b"MZ") or head.startswith(MACHO_MAGICS):
+        return BinaryArtifact(path=path, kind="executable_binary", detail="executable binary header")
+    if suffix in KNOWN_BINARY_SUFFIXES:
+        return BinaryArtifact(path=path, kind="binary_blob", detail="binary extension")
+    if b"\x00" in head:
+        return BinaryArtifact(path=path, kind="binary_blob", detail="contains NUL bytes")
+    return None
+
+
+def iter_text_files(
+    root: Path,
+    max_files: int,
+    max_bytes: int,
+    max_binary_artifacts: int,
+    max_binary_bytes: int,
+) -> FileInventory:
+    files: list[Path] = []
+    binary_artifacts: list[BinaryArtifact] = []
+    total_bytes = 0
+    total_files = 0
+    total_binary_bytes = 0
+    for path in root.rglob("*"):
+        if not path.is_file():
+            continue
+        try:
+            size = path.stat().st_size
+        except OSError:
+            continue
+        total_files += 1
+        total_bytes += size
+        if total_bytes > max_bytes:
+            raise ScanError("Scan size exceeded max_bytes policy limit")
+        if total_files > max_files:
+            raise ScanError("Scan file count exceeded max_files policy limit")
+        classification = _classify_non_text(path)
+        if classification is not None:
+            binary_artifacts.append(classification)
+            total_binary_bytes += size
+            if len(binary_artifacts) > max_binary_artifacts:
+                raise ScanError("Binary artifact count exceeded max_binary_artifacts policy limit")
+            if total_binary_bytes > max_binary_bytes:
+                raise ScanError("Binary artifact bytes exceeded max_binary_bytes policy limit")
+            continue
+        files.append(path)
+    return FileInventory(text_files=files, binary_artifacts=binary_artifacts)
+
+
+def _safe_read_text(path: Path) -> str:
+    try:
+        raw = path.read_bytes()
+    except OSError:
+        return ""
+    if b"\x00" in raw:
+        return ""
+    return raw.decode("utf-8", errors="ignore")
+
+
+def _normalize_text(text: str) -> str:
+    norm = unicodedata.normalize("NFKC", text)
+    norm = ZERO_WIDTH_RE.sub("", norm)
+    # Recover common defanged URL/domain forms.
+    norm = re.sub(r"\bhxxps://", "https://", norm, flags=re.IGNORECASE)
+    norm = re.sub(r"\bhxxp://", "http://", norm, flags=re.IGNORECASE)
+    norm = norm.replace("[.]", ".").replace("(.)", ".").replace("{.}", ".")
+    return norm
+
+
+def _try_decode_b64(token: str) -> str | None:
+    value = token.strip()
+    if len(value) < 16:
+        return None
+    # Normalize urlsafe base64 and missing padding.
+    value = value.replace("-", "+").replace("_", "/")
+    padding = (-len(value)) % 4
+    if padding:
+        value += "=" * padding
+    try:
+        blob = base64.b64decode(value, validate=True)
+    except (binascii.Error, ValueError):
+        return None
+    if not blob:
+        return None
+    candidate = blob.decode("utf-8", errors="ignore").strip()
+    if not candidate:
+        return None
+    printable_ratio = sum(ch.isprintable() for ch in candidate) / max(len(candidate), 1)
+    if printable_ratio < 0.9:
+        return None
+    return candidate
+
+
+def _decode_base64_fragments(text: str, max_decodes: int = 6) -> list[str]:
+    decoded: list[str] = []
+    seen: set[str] = set()
+    for token in B64_TOKEN_RE.findall(text):
+        if token in seen:
+            continue
+        seen.add(token)
+        if len(decoded) >= max_decodes:
+            break
+        candidate = _try_decode_b64(token)
+        if candidate:
+            decoded.append(candidate)
+
+    # Also try concatenated quoted fragments often used to bypass token scanners.
+    for line in text.splitlines():
+        fragments = QUOTED_B64_FRAGMENT_RE.findall(line)
+        if len(fragments) < 2:
+            continue
+        joined = "".join(fragments)
+        if joined in seen:
+            continue
+        seen.add(joined)
+        if len(decoded) >= max_decodes:
+            break
+        candidate = _try_decode_b64(joined)
+        if candidate:
+            decoded.append(candidate)
+    return decoded
+
+
+def _prepare_analysis_text(text: str) -> str:
+    norm = _normalize_text(text)
+    decoded = _decode_base64_fragments(norm)
+    if not decoded:
+        return norm
+    return f"{norm}\n\n# decoded_fragments\n" + "\n".join(decoded)
+
+
+def _extract_actions(text: str, action_patterns: dict[str, re.Pattern[str]]) -> set[str]:
+    """Return the set of action names whose patterns match anywhere in *text*.
+
+    Used for whole-file matching (backward-compatible). For chain rule detection
+    use :func:`_extract_actions_windowed` which applies a proximity constraint.
+    """
+    actions: set[str] = set()
+    for action, pattern in action_patterns.items():
+        if pattern.search(text):
+            actions.add(action)
+    return actions
+
+
+_LANG_EXTENSIONS: dict[str, frozenset[str]] = {
+    "javascript": frozenset({".js", ".ts", ".mjs", ".cjs"}),
+    "ruby": frozenset({".rb"}),
+    "go": frozenset({".go"}),
+    "rust": frozenset({".rs"}),
+    "python": frozenset({".py"}),
+    "shell": frozenset({".sh", ".bash", ".zsh"}),
+}
+_CHAIN_WINDOW_LINES = 40
+
+
+def _extract_actions_windowed(
+    text: str,
+    action_patterns: dict[str, re.Pattern[str]],
+    window_lines: int = _CHAIN_WINDOW_LINES,
+) -> list[set[str]]:
+    """Return a list of action sets, one per sliding window of *window_lines* lines.
+
+    For files shorter than *window_lines* a single whole-file window is returned,
+    preserving the original behaviour for small skill files.
+    """
+    lines = text.splitlines()
+    if len(lines) <= window_lines:
+        return [_extract_actions(text, action_patterns)]
+    windows: list[set[str]] = []
+    step = max(1, window_lines // 2)  # 50 % overlap
+    for start in range(0, len(lines), step):
+        chunk = "\n".join(lines[start : start + window_lines])
+        window_actions = _extract_actions(chunk, action_patterns)
+        if window_actions:
+            windows.append(window_actions)
+    return windows if windows else [set()]
+
+
+def _normalize_domain(value: str) -> str:
+    return value.lower().strip().strip(".,;)")
+
+
+def _parse_requirements(text: str) -> list[tuple[str, str]]:
+    deps: list[tuple[str, str]] = []
+    for line in text.splitlines():
+        line = line.strip()
+        if not line or line.startswith("#"):
+            continue
+        if "==" in line:
+            name, version = line.split("==", 1)
+            deps.append((name.strip().lower(), version.strip()))
+    return deps
+
+
+def _find_unpinned_requirements(text: str) -> list[tuple[str, str]]:
+    unpinned: list[tuple[str, str]] = []
+    for line in text.splitlines():
+        line = line.strip()
+        if not line or line.startswith("#"):
+            continue
+        if "==" in line:
+            continue
+        if any(op in line for op in (">=", "<=", "~=", ">", "<")):
+            unpinned.append((line.split()[0], line))
+    return unpinned
+
+
+def _parse_package_json(text: str) -> list[tuple[str, str]]:
+    deps: list[tuple[str, str]] = []
+    try:
+        data = json.loads(text)
+    except json.JSONDecodeError:
+        return deps
+    for section in ("dependencies", "devDependencies"):
+        values = data.get(section, {})
+        if isinstance(values, dict):
+            for name, version in values.items():
+                deps.append((name.lower(), str(version)))
+    return deps
+
+
+def _parse_package_scripts(text: str) -> dict[str, str]:
+    try:
+        data = json.loads(text)
+    except json.JSONDecodeError:
+        return {}
+    scripts = data.get("scripts", {})
+    if not isinstance(scripts, dict):
+        return {}
+    parsed: dict[str, str] = {}
+    for name, cmd in scripts.items():
+        if isinstance(name, str) and isinstance(cmd, str):
+            parsed[name.lower()] = cmd
+    return parsed
+
+
+def _is_unpinned_npm(version: str) -> bool:
+    v = version.strip().lower()
+    if not v or v == "latest":
+        return True
+    return v.startswith("^") or v.startswith("~") or "*" in v or "x" in v
+
+
+def _extract_iocs(path: Path, text: str) -> list[IOC]:
+    iocs: list[IOC] = []
+    seen: set[tuple[str, str]] = set()
+    for url in URL_RE.findall(text):
+        key = ("url", url)
+        if key not in seen:
+            seen.add(key)
+            iocs.append(IOC(value=url, kind="url", source_path=str(path)))
+    for token in IP_RE.findall(text):
+        try:
+            ipaddress.ip_address(token)
+        except ValueError:
+            continue
+        key = ("ip", token)
+        if key not in seen:
+            seen.add(key)
+            iocs.append(IOC(value=token, kind="ip", source_path=str(path)))
+    for domain in DOMAIN_RE.findall(text):
+        norm = _normalize_domain(domain)
+        if norm.startswith("http"):
+            continue
+        tld = norm.rsplit(".", 1)[-1]
+        if tld not in KNOWN_TLDS:
+            continue
+        key = ("domain", norm)
+        if key not in seen:
+            seen.add(key)
+            iocs.append(IOC(value=norm, kind="domain", source_path=str(path)))
+    return iocs
+
+
+def _load_builtin_vuln_db() -> VulnDB:
+    raw = resources.files("skillscan.data.intel").joinpath("vuln_db.json").read_text(encoding="utf-8")
+    return cast(VulnDB, json.loads(raw))
+
+
+def _load_builtin_ioc_db() -> IOCDB:
+    raw = resources.files("skillscan.data.intel").joinpath("ioc_db.json").read_text(encoding="utf-8")
+    return cast(IOCDB, json.loads(raw))
+
+
+def _merge_user_intel(ioc_db: IOCDB, vuln_db: VulnDB) -> tuple[IOCDB, VulnDB, list[str]]:
+    store = load_store()
+    sources = ["builtin:ioc_db", "builtin:vuln_db"]
+    for source in store.sources:
+        if not source.enabled:
+            continue
+        path = Path(source.path)
+        if not path.exists():
+            continue
+        try:
+            payload = json.loads(path.read_text(encoding="utf-8"))
+        except (json.JSONDecodeError, OSError, UnicodeDecodeError):
+            continue
+        if source.kind == "ioc" and isinstance(payload, dict):
+            for key in ("domains", "ips", "urls", "cidrs"):
+                values = payload.get(key, [])
+                if isinstance(values, list):
+                    ioc_db.setdefault(key, []).extend([str(v).lower() for v in values])
+        if source.kind == "vuln" and isinstance(payload, dict):
+            for ecosystem, values in payload.items():
+                eco = vuln_db.setdefault(ecosystem, {})
+                if isinstance(values, dict):
+                    eco.update(values)
+        sources.append(f"user:{source.name}")
+    for key in ("domains", "ips", "urls", "cidrs"):
+        deduped = sorted(set(ioc_db.get(key, [])))
+        ioc_db[key] = deduped
+    return ioc_db, vuln_db, sources
+
+
+def _binary_kind_template(kind: str) -> tuple[str, Severity, str, str]:
+    if kind == "executable_binary":
+        return (
+            "BIN-001",
+            Severity.HIGH,
+            "Executable binary artifact present",
+            "Review and verify binary provenance and signatures before trust or execution.",
+        )
+    if kind == "binary_library":
+        return (
+            "BIN-002",
+            Severity.MEDIUM,
+            "Compiled library artifact present",
+            "Verify shared/static library provenance and hash against trusted release artifacts.",
+        )
+    if kind == "binary_blob":
+        return (
+            "BIN-003",
+            Severity.MEDIUM,
+            "Binary blob artifact present",
+            "Inspect binary payload purpose and source; prefer transparent source artifacts when possible.",
+        )
+    return (
+        "BIN-004",
+        Severity.LOW,
+        "Compiled Python bytecode artifact present",
+        (
+            "Prefer scanning source files with bytecode provenance; "
+            "review for hidden or mismatched source content."
+        ),
+    )
+
+
+def _binary_artifact_findings(artifacts: list[BinaryArtifact]) -> list[Finding]:
+    grouped: dict[str, list[BinaryArtifact]] = {}
+    for item in artifacts:
+        grouped.setdefault(item.kind, []).append(item)
+    findings: list[Finding] = []
+    for kind, items in grouped.items():
+        rule_id, severity, title, mitigation = _binary_kind_template(kind)
+        sample = ", ".join(str(item.path) for item in items[:3])
+        findings.append(
+            Finding(
+                id=rule_id,
+                category="binary_artifact",
+                severity=severity,
+                confidence=0.9,
+                title=title,
+                evidence_path=str(items[0].path),
+                snippet=f"count={len(items)} sample={sample}",
+                mitigation=mitigation,
+            )
+        )
+    return findings
+
+
+def _ip_in_cidrs(ip_value: str, cidrs: list[str]) -> bool:
+    try:
+        ip_obj = ipaddress.ip_address(ip_value)
+    except ValueError:
+        return False
+    for cidr in cidrs:
+        try:
+            if ip_obj in ipaddress.ip_network(cidr, strict=False):
+                return True
+        except ValueError:
+            continue
+    return False
